@@ -18,25 +18,25 @@
   "author": "Pavel Bychko <abordage.dev@gmail.com> (https://abordage.dev)",
   "license": "MIT",
   "dependencies": {
-    "@actions/core": "^1.9.1",
-    "@octokit/rest": "^18.12.0",
+    "@actions/core": "^1.10.0",
+    "@octokit/rest": "^19.0.7",
     "chrome-launcher": "^0.15.1",
-    "dotenv": "^16.0.1",
-    "lighthouse": "^9.6.1"
+    "dotenv": "^16.0.3",
+    "lighthouse": "^9.6.8"
   },
   "devDependencies": {
-    "@types/node": "^17.0.36",
-    "@types/node-fetch": "^2.6.1",
-    "@typescript-eslint/eslint-plugin": "^5.27.0",
-    "@typescript-eslint/parser": "^5.27.0",
-    "@vercel/ncc": "^0.34.0",
-    "eslint": "^8.16.0",
+    "@types/node": "^18.14.1",
+    "@types/node-fetch": "2.6.0",
+    "@typescript-eslint/eslint-plugin": "^5.53.0",
+    "@typescript-eslint/parser": "^5.53.0",
+    "@vercel/ncc": "^0.36.1",
+    "eslint": "^8.34.0",
     "eslint-config-standard": "^17.0.0",
-    "eslint-plugin-import": "^2.26.0",
-    "eslint-plugin-n": "^15.2.1",
-    "eslint-plugin-promise": "^6.0.0",
-    "ts-node": "^10.8.0",
-    "tslib": "^2.4.0",
-    "typescript": "^4.7.2"
+    "eslint-plugin-import": "^2.27.5",
+    "eslint-plugin-n": "^15.6.1",
+    "eslint-plugin-promise": "^6.1.1",
+    "ts-node": "^10.9.1",
+    "tslib": "^2.5.0",
+    "typescript": "^4.9.5"
   }
 }
