@@ -2,8 +2,9 @@ import * as core from '@actions/core';
 import {Octokit} from '@octokit/rest';
 import {config} from 'dotenv';
 import {resolve} from 'path';
-const chromeLauncher = require('chrome-launcher');
-const lighthouse = require('lighthouse');
+import * as chromeLauncher from 'chrome-launcher';
+import * as lighthouse from 'lighthouse';
+import {Flags} from 'lighthouse/types/externs';
 
 config({path: resolve(__dirname, '../.env')});
 
@@ -20,46 +21,40 @@ const summaryTable = [];
 const title = 'My website [update ' + updateDate + ']';
 
 (async () => {
-  /**
-   * Get metrics
-   */
+  /** Get metrics */
   const chrome = await chromeLauncher.launch({chromeFlags: ['--headless']});
-  const options = {logLevel: 'info', output: 'json', port: chrome.port};
+  const options: Flags = {logLevel: 'info', output: 'json', port: chrome.port};
   const runnerResult = await lighthouse(TEST_URL, options);
   await chrome.kill();
 
   summaryTable.push([{data: 'Category', header: true}, {data: 'Result', header: true}]);
 
-  /**
-   * Formatting
-   */
+  /** Formatting */
   const gistContent =
-    [
-      ['Performance', runnerResult.lhr.categories.performance.score * 100],
-      ['Accessibility', runnerResult.lhr.categories.accessibility.score * 100],
-      ['Best Practices', runnerResult.lhr.categories['best-practices'].score * 100],
-      ['SEO', runnerResult.lhr.categories.seo.score * 100],
-      ['PWA Ready', runnerResult.lhr.categories.pwa.score * 100]
-    ]
-      .map((content) => {
-        summaryTable.push([content[0], content[1] + '%']);
+        [
+          ['Performance', runnerResult.lhr.categories.performance.score * 100],
+          ['Accessibility', runnerResult.lhr.categories.accessibility.score * 100],
+          ['Best Practices', runnerResult.lhr.categories['best-practices'].score * 100],
+          ['SEO', runnerResult.lhr.categories.seo.score * 100],
+          ['PWA Ready', runnerResult.lhr.categories.pwa.score * 100]
+        ]
+          .map((content) => {
+            summaryTable.push([content[0], content[1] + '%']);
 
-        let badge = '🙉';
-        if (content[1] > 80) badge = '🥈';
-        if (content[1] > 90) badge = '🥇';
-        if (content[1] === 100) badge = '🏆';
+            let badge = '🙉';
+            if (content[1] > 80) badge = '🥈';
+            if (content[1] > 90) badge = '🥇';
+            if (content[1] === 100) badge = '🏆';
 
-        const title = (content[0] + ':').padEnd(RESULT_BADGE ? 37 : 49, '.');
-        const percent = (content[1] + '%').padStart(4, '.');
-        const result = RESULT_BADGE ? ' ' + (' ' + badge).padStart(11, '.') : '';
+            const title = (content[0] + ':').padEnd(RESULT_BADGE ? 42 : 54, '.');
+            const percent = (content[1] + '%').padStart(4, '.');
+            const result = RESULT_BADGE ? ' ' + (' ' + badge).padStart(11, '.') : '';
 
-        return title + percent + result;
-      })
-      .join('\n');
+            return title + percent + result;
+          })
+          .join('\n');
 
-  /**
-   * Get gist filename
-   */
+  /** Get gist filename */
   const octokit = new Octokit({auth: GH_TOKEN});
   const gist = await octokit.gists.get({gist_id: GIST_ID})
     .catch(error => core.setFailed('Action failed: Gist ' + error.message));
@@ -72,8 +67,8 @@ const title = 'My website [update ' + updateDate + ']';
   }
 
   /**
-   * Update gist
-   */
+     * Update gist
+     */
   await octokit.gists.update({
     gist_id: GIST_ID,
     files: {
@@ -85,14 +80,14 @@ const title = 'My website [update ' + updateDate + ']';
   }).catch(error => core.setFailed('Action failed: Gist ' + error.message));
 
   /**
-   * Print summary
-   */
+     * Print summary
+     */
   const summary = core.summary
     .addHeading('Results')
     .addTable(summaryTable)
     .addBreak()
     .addRaw('Lighthouse metrics for ')
-    .addLink(runnerResult.lhr.finalUrl, runnerResult.lhr.finalUrl)
+    .addLink(runnerResult.lhr.mainDocumentUrl, runnerResult.lhr.mainDocumentUrl)
     .addRaw(' generated by ')
     .addLink('lighthouse-box/1.0', ACTION_URL);
 
